3. PostController로 전환
